refactor: admin/mirai-stance featureにrepositoryレイヤーを追加
