+++ admin/src/features/mirai-stance/actions/create-stance.ts
@@ -1,23 +1,12 @@
 "use server";
 
-import { createAdminClient } from "@mirai-gikai/supabase";
 import { invalidateWebCache } from "@/lib/utils/cache-invalidation";
 import type { StanceInput } from "../types";
+import { createMiraiStance } from "../repositories/mirai-stance-repository";
 
 export async function createStance(billId: string, data: StanceInput) {
   try {
-    const supabase = createAdminClient();
-
-    const { error } = await supabase.from("mirai_stances").insert({
-      bill_id: billId,
-      type: data.type,
-      comment: data.comment || null,
-    });
-
-    if (error) {
-      console.error("Error creating stance:", error);
-      throw new Error("スタンスの作成に失敗しました");
-    }
+    await createMiraiStance(billId, data);
 
     invalidateWebCache();
     return { success: true };
